+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Change Log
 
 - Added API property `Renderer.background`
 - Added camera controller `Bu.MouseControl`
-
+- Added `pixelWidth` and `pixelHeight` to `Renderer`
 
 ## 0.4.0 / 2016/11/15